refactor: Constants 인터페이스 구현체로 전부 변경
일관성을 높일 수 있다
