+++ src/main/java/lotto/constants/Constants.java
@@ -1,0 +1,7 @@
+package lotto.constants;
+
+public interface Constants<T> {
+
+    T getInstance();
+
+}
