+++ src/main/java/lotto/constants/message/InputError.java
@@ -1,6 +1,8 @@
 package lotto.constants.message;
 
-public enum InputError implements Message {
+import lotto.constants.Constants;
+
+public enum InputError implements Constants<String> {
     DUPLICATE_LOTTO_NUMBER("[ERROR] 로또의 번호 조합에 중복이 숫자가 발견되었습니다."),
     DUPLICATE_BONUS_NUMBER("[ERROR] 보너스 숫자가 당첨 번호 안에 이미 존재합니다.");
 
@@ -11,7 +13,7 @@ public enum InputError implements Message {
     }
 
     @Override
-    public String getMessage() {
+    public String getInstance() {
         return this.message;
     }
 }
