--- src/main/java/lotto/constants/message/Message.java
@@ -1,7 +1,0 @@
-package lotto.constants.message;
-
-public interface Message {
-
-    String getMessage();
-
-}
