+++ src/main/java/lotto/constants/message/RangeError.java
@@ -1,10 +1,11 @@
 package lotto.constants.message;
 
+import lotto.constants.Constants;
 import lotto.constants.value.LottoRule;
 
-public enum RangeError implements Message {
-    NUMBER("[ERROR] 로또 번호는 "+ LottoRule.MINIMUM_NUMBER_RANGE.getValue() +"부터 "+ LottoRule.MAXIMUM_NUMBER_RANGE.getValue() +"사이여야 합니다."),
-    LOTTO("[ERROR] 로또의 길이가 "+LottoRule.COMBINATION_LENGTH.getValue()+"과 일치하지 않습니다");
+public enum RangeError implements Constants<String> {
+    NUMBER("[ERROR] 로또 번호는 "+ LottoRule.MINIMUM_NUMBER_RANGE.getInstance() +"부터 "+ LottoRule.MAXIMUM_NUMBER_RANGE.getInstance() +"사이여야 합니다."),
+    LOTTO("[ERROR] 로또의 길이가 "+LottoRule.COMBINATION_LENGTH.getInstance()+"과 일치하지 않습니다");
 
     private String message;
 
@@ -13,7 +14,7 @@ public enum RangeError implements Message {
     }
 
     @Override
-    public String getMessage() {
+    public String getInstance() {
         return message;
     }
 }
