+++ src/main/java/lotto/domain/ComponentNumber.java
@@ -17,9 +17,9 @@ public class ComponentNumber implements Component {
     }
 
     private static void validateRange(int lottoNumber) {
-        if (lottoNumber < LottoRule.MINIMUM_NUMBER_RANGE.getValue() || lottoNumber > MAXIMUM_NUMBER_RANGE.getValue()) {
-            System.out.println(RangeError.NUMBER.getMessage());
-            throw new IllegalArgumentException(RangeError.NUMBER.getMessage());
+        if (lottoNumber < LottoRule.MINIMUM_NUMBER_RANGE.getInstance() || lottoNumber > MAXIMUM_NUMBER_RANGE.getInstance()) {
+            System.out.println(RangeError.NUMBER.getInstance());
+            throw new IllegalArgumentException(RangeError.NUMBER.getInstance());
         }
     }
