+++ src/main/java/lotto/domain/LottoMachine.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.constants.ScoreSystem;
+import lotto.constants.collection.ScoreSystem;
 import lotto.domain.factory.LottoFactory;
 import lotto.domain.score.Score;
 import lotto.dto.ScoreDto;
