+++ src/main/java/lotto/domain/score/Score.java
@@ -1,6 +1,6 @@
 package lotto.domain.score;
 
-import lotto.constants.ScoreSystem;
+import lotto.constants.collection.ScoreSystem;
 import lotto.domain.BonusComponent;
 import lotto.domain.Component;
 import lotto.domain.Lotto;
